Implement Entity and DTO classes; Fix SQL

+++ liquibase/changelogs/homework-2/sql/05.create-comments-table.sql
@@ -1,8 +1,8 @@
 CREATE TABLE comments
 (
     id           SERIAL PRIMARY KEY,
-    task_id      INT,
-    employee_id  INT,
+    task_id      BIGINT,
+    employee_id  BIGINT,
     comment_text TEXT NOT NULL,
     created_at   TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
     FOREIGN KEY (task_id) REFERENCES tasks (id),
